% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{Beta}
\alias{Beta}
\title{Calculate Beta}
\usage{
Beta(kD, S)
}
\arguments{
\item{kD}{Hydraulic conductivity (L2/T)}

\item{S}{Storage coefficient (-)}
}
\value{
Beta (T-1)
}
\description{
Calculate Beta
}
