% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1_one_dimensional_flow.R
\name{Br_123_32}
\alias{Br_123_32}
\title{Bruggeman Formula 123.32, p. 66}
\usage{
Br_123_32(x, t, S, kD, c, h)
}
\arguments{
\item{x}{Distance (L)}

\item{t}{Time (T)}

\item{S}{Storage coefficient (-)}

\item{kD}{Hydraulic conductivity (L2/T)}

\item{c}{Hydraulic resistance (T)}

\item{h}{Constant in the definition of the stress (= sudden drawdown) (L)}
}
\value{
Phi Head change (L)
}
\description{
BI. One-dimensional groundwater flow
BI-2. One-dimensional groundwater flow in a semi-infinite field
The soil is assumed to be homogeneous
The flow is non-periodic
The boundary condition at x -- 0 a is given head or drawdown
Leaky aquifers with variable head at x = 0
Sudden drawdown of the surface water level, which is kept constant thereafter.
}
\details{
Calculate the Head change Phi at distance x and time t due to a stress
change h (= sudden drawdown) at t=0 at x=0.
}
