% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1_one_dimensional_flow.R
\name{Br_123_34}
\alias{Br_123_34}
\title{Formula 123.34, p. 67.}
\usage{
Br_123_34(x, t, S, kD, c, a)
}
\arguments{
\item{x}{Distance (L)}

\item{t}{Time (T)}

\item{S}{Storage coefficient (-)}

\item{kD}{Hydraulic conductivity (L2/T)}

\item{c}{Hydraulic resistance (T)}

\item{a}{Constant in the linear drawdown (F(t)=at) (L/T)}
}
\value{
Phi Head change (L)
}
\description{
Calculate the head change Phi at distance x and time t due to a stress
change F(t)=at: linear drawdown of the surface water level at x=0.
}
\section{Schematization}{


\itemize{
\item BI. One-dimensional groundwater flow
\item BI-2. One-dimensional groundwater flow in a semi-infinite field
\item The soil is assumed to be homogeneous
\item The flow is non-periodic
\item The boundary condition at x = 0 a is given head or drawdown
\item Leaky aquifers with variable head at x = 0
}
}

