% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsolu.r
\name{brsolu}
\alias{brsolu}
\title{Analytical Solutions of Geohydrological Problems.}
\description{
Functions as defined in Part A ("Mathematical tools") in the book
"Analytical solutions of geohydrological problems" by G.A. Bruggeman.
}
\section{Chapter B1}{
 One-dimensional grondwater flow:

\code{\link{Br_123_32}}

\code{\link{d_Br_123_32_dx}}
}

