% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1_one_dimensional_flow.R
\name{d_Br_123_32_dx}
\alias{d_Br_123_32_dx}
\title{Bruggeman Formula 123.32 p. 66
First Derivative of Phi with respect to x}
\usage{
d_Br_123_32_dx(x, t, S, kD, c, h)
}
\arguments{
\item{x}{Distance (L)}

\item{t}{Time (T)}

\item{S}{Storage coefficient (-)}

\item{kD}{Hydraulic conductivity (L2/T)}

\item{c}{Hydraulic resistance (T)}

\item{h}{Constant in the definition of the stress (= sudden drawdown) (L)}
}
\value{
First Derivative of Phi with respect to x (-)
}
\description{
Bruggeman Formula 123.32 p. 66
First Derivative of Phi with respect to x
}
