% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{Labda}
\alias{Labda}
\title{Calculate the Leakage factor}
\usage{
Labda(c, kD)
}
\arguments{
\item{c}{Hydraulic resistance (T)}

\item{kD}{Hydraulic conductivity (L2/T)}
}
\value{
Leakage factor (L)
}
\description{
Calculate the Leakage factor
}
